Working POC! strip ICC profile + minor refactor
- move extractIccFromJpeg() into it's own JpegUtils class
- adjust extraction to return a data object containing the start and end position of the APP2 section so the caller can know where to strip bytes.
- function to strip bytes
